% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{adjcentScOfSP}
\alias{adjcentScOfSP}
\title{adjcentScOfSP}
\usage{
adjcentScOfSP(obj)
}
\arguments{
\item{obj}{Integrated Seurat object.}
}
\value{
A matrix of distance weights between ST and SC.
}
\description{
Generating distance matrix between ST spots and SC cells.
}
