% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightDist.R
\name{adjcentScOfSPGlobal}
\alias{adjcentScOfSPGlobal}
\title{adjcentScOfSPGlobal}
\usage{
adjcentScOfSPGlobal(obj)
}
\arguments{
\item{obj}{Integrated Seurat object.}
}
\value{
A matrix of distance weights between ST and SC.
}
\description{
Generating distance matrix between ST spots and SC cells.
}
