% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{adjustNetProb}
\alias{adjustNetProb}
\title{adjustNetProb}
\usage{
adjustNetProb(netx.pred, sc.obj, sp.obj, hot.pvals, p.cut = 0.05)
}
\arguments{
\item{netx.pred}{Matrix or data frame of network predictions.}

\item{sc.obj}{Seurat object of SC data.}

\item{sp.obj}{Seurat object of ST data.}

\item{hot.pvals}{A data frame of hotspot p-values indicating hotspot presence in spots and cell types.}

\item{p.cut}{Numeric p-value cutoff to call hotspots. Default: 0.05.}
}
\value{
Adjusted network predictions matrix where weights are applied based on cell type proportions and hotspot information.
}
\description{
Adjusts FNN predictions based on cell type proportions derived from SC and ST data, considering hotspot information.
}
