% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estCellPerSpots.R
\name{adjustScObj}
\alias{adjustScObj}
\title{adjustScObj}
\usage{
adjustScObj(sc.obj, st.prop.cnts, assay = "SCT")
}
\arguments{
\item{sc.obj}{Seurat object of single-cell (SC) data.}

\item{assay}{Assay type to use for SC data normalization. Default: SCT.}

\item{st.prop}{A data.frame containing cell counts for ST data.}
}
\value{
Adjusted Seurat object of SC data.
}
\description{
Adjust single-cell (SC) data by incorporating inferred proportions and cell counts for mapping to spatial transcriptomics (ST) spots
}
