% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{assignSc2SP}
\alias{assignSc2SP}
\title{assignSc2SP}
\usage{
assignSc2SP(sp.obj, sc.obj, out.sim, num.cells, st.pvals, duplicated = FALSE)
}
\arguments{
\item{sp.obj}{Seurat object of ST data.}

\item{sc.obj}{Seurat object of SC data.}

\item{out.sim}{Similarity matrix between single-cells and spots.}

\item{num.cells}{Number of cells projected to spots. Default: 10.}

\item{st.pvals}{A data.frame of p-values of ST spots.}

\item{duplicated}{Unique assign single-cells to spot or not. Default: FALSE.}
}
\value{
A list of assigned cells corresponding to spots.
}
\description{
Assign single-cells to spots using linear assignment method.
}
