% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{binarizeMax}
\alias{binarizeMax}
\title{binarizeMax}
\usage{
binarizeMax(sp.score)
}
\arguments{
\item{sp.score}{A matrix of scores (e.g., spatial transcriptomics scores). Each row represents a different spot or entity, and each column represents a different gene or feature.}
}
\value{
A binary matrix of the same size as the input matrix, where each row contains a 1 at the position of the maximum value in that row, and 0 elsewhere.
}
\description{
Converts a matrix of scores into a binary matrix where each row contains a 1 at the position of the maximum value in that row.
}
