% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{calcSimlarityDist}
\alias{calcSimlarityDist}
\title{calcSimilarityDist}
\usage{
calcSimlarityDist(
  sc.obj,
  sp.obj,
  sc.score,
  sp.score,
  feature.based,
  sc.markers
)
}
\arguments{
\item{sc.obj}{Seurat object containing SC data (assay slot must hold a counts/data matrix).}

\item{sp.obj}{Seurat object containing ST data.}

\item{sc.score}{Numeric matrix/data frame of SC scores (\strong{cells × features}).}

\item{sp.score}{Numeric matrix/data frame of ST scores (\strong{spots × features}).}

\item{feature.based}{Specify whether features for likelihood or correlation calculations between single cells and spots are based on gene expression ('gene.based') or signature scores of cell types ('celltype.based').}

\item{sc.markers}{List of marker genes per cell type.}
}
\value{
A numeric matrix of shape \strong{spots × cells} with spot–cell similarities.
}
\description{
Builds a spot–cell similarity matrix.
}
