% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region_specific_gene.R
\name{calculate_cohens_d}
\alias{calculate_cohens_d}
\title{Compare Gene Expression Between Regions Using t-test}
\usage{
calculate_cohens_d(x, y)
}
\arguments{
\item{x}{Numeric vector for group 1}

\item{y}{Numeric vector for group 2}
}
\value{
Numeric value representing Cohen's d
}
\description{
Compare Gene Expression Between Regions Using t-test
}
\keyword{internal}
