% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Region_specific_gene.R
\name{calculate_rank_biserial_correlation}
\alias{calculate_rank_biserial_correlation}
\title{Calculate rank-biserial correlation as effect size for Wilcoxon test}
\usage{
calculate_rank_biserial_correlation(x, y)
}
\arguments{
\item{x}{Numeric vector for group 1}

\item{y}{Numeric vector for group 2}
}
\value{
Numeric value representing the rank-biserial correlation
}
\description{
Calculate rank-biserial correlation as effect size for Wilcoxon test
}
\keyword{internal}
