% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkParams.R
\name{checkParams.runMap2SP}
\alias{checkParams.runMap2SP}
\title{checkParams}
\usage{
checkParams.runMap2SP(argg)
}
\arguments{
\item{argg}{A list of input arguments.}
}
\value{
TRUE
}
\description{
Check the legality of input arguments.
}
