% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coExistIndex}
\alias{coExistIndex}
\title{coExistIndex}
\usage{
coExistIndex(sce, min.cells = 0)
}
\arguments{
\item{sce}{Seurat or SingleCellExperiment object returned from runMap2Sp function.}

\item{min.cells}{Cell types with number of cells must be greater than k. Default: 0.}
}
\value{
A data.frame of coexists J-index.
}
\description{
Calculation of co-exists index.
}
