% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createSpatialObject}
\alias{createSpatialObject}
\title{createSpatialObject}
\usage{
createSpatialObject(
  counts,
  coord.df,
  coord.label = c("x", "y"),
  meta.data = NULL,
  class = "SlideSeq"
)
}
\arguments{
\item{counts}{UMI count matrix, rows are genes and columns are barcodes.}

\item{coord.df}{A data.frame of coordinates for spatial barcodes. The coordinates are placed in the first two columns. Default: c("x", "y").}

\item{meta.data}{A data frame containing metadata to be added to the Seurat object. Default: NULL.}

\item{class}{Class name of the Spatial image slot. Default: SlideSeq.}
}
\value{
A Seurat object for ST data.
}
\description{
Create Spatial seurat object for ST data from other platforms.
}
