% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv.R
\name{ctypesOfClusters}
\alias{ctypesOfClusters}
\title{ctypesOfClusters}
\usage{
ctypesOfClusters(sp.obj, hot.pvals, p.cut = 0.05, w.cut = 0.01)
}
\arguments{
\item{sp.obj}{Seurat object containing spatial transcriptomics (ST) data.}

\item{hot.pvals}{A data frame containing p-values from the hotspot detection step. Each row corresponds to a spatial spot, and columns represent different hotspot detection metrics.}

\item{p.cut}{Threshold for p-values to filter significant hotspot areas. Spots with p-values below this threshold are considered significant. Default: 0.05.}

\item{w.cut}{Threshold for the weighted score. Cellular types with scores below this threshold are set to 0, and those above are set to 1. Default: 0.01.}
}
\value{
A data frame where each row corresponds to a cluster and each column represents a cellular type, with values indicating weighted scores based on the presence of significant hotspots.
}
\description{
Calculates weighted scores of cellular types across clusters based on spatial transcriptomics (ST) data and hotspot detection results. The function evaluates the significance of hotspot areas and assigns weighted scores to cellular types in each cluster.
}
