% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectHotSpots.R
\name{dectHotSpotsByGetisOrdGi}
\alias{dectHotSpotsByGetisOrdGi}
\title{dectHotSpotsByGetisOrdGi}
\usage{
dectHotSpotsByGetisOrdGi(sp.obj, sp.score, knn = 5, p.value.threshold = 0.05)
}
\arguments{
\item{sp.obj}{Seurat object of spatial transcriptome (ST) data.}

\item{sp.score}{Signature score of cell types.}

\item{knn}{K nearest cells used. Default: 5.}

\item{p.value.threshold}{P-value threadhold for filtering out hot-spots. Default: 0.05.}
}
\value{
A data frame containing hotspot information. A value of TRUE indicates a detected hotspot, while any other value indicates a non-hotspot.
}
\description{
Detect hot spots using Getis-Ord Gi* index
}
