% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{downSamplSeurat}
\alias{downSamplSeurat}
\title{downSamplSeurat}
\usage{
downSamplSeurat(obj.seu, cnt = 500, percent = NULL, seed = 123456)
}
\arguments{
\item{obj.seu}{Seurat object}

\item{cnt}{Sampling size for each ident. Default: 500.}

\item{percent}{Sampling percentage of each ident (value range: 0-1). Default: NULL.}

\item{seed}{Set random seed, default: 123456.}
}
\value{
Subset of seurat object.
}
\description{
Down-sampling for Seurat obeject.
}
