% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estCellPerSpots.R
\name{estCellPerSpots}
\alias{estCellPerSpots}
\title{estCellPerSpots}
\usage{
estCellPerSpots(sp.obj, max.cells = 10, quantile.cut = 0.99, num.genes = 2000)
}
\arguments{
\item{sp.obj}{Seurat object of spatial transcriptome (ST) data.}

\item{max.cells}{Maximum number of cells in Spots. Default: 10 (10x Genomics).}
}
\value{
A vector of estimated cells per spot.
}
\description{
Estimate number of cells per spot.
}
