% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{estPvalsOfSpot}
\alias{estPvalsOfSpot}
\title{estPvalsOfSpot}
\usage{
estPvalsOfSpot(sp.obj, sp.score, knn = 5)
}
\arguments{
\item{sp.obj}{Seurat object of spatial transcriptome (ST) data.}

\item{sp.score}{Signature score of cell types.}

\item{knn}{K nearest cells used. Default: 5.}
}
\value{
A data.frame of p-values.
}
\description{
Infer whether signature score in spot significant higher than background.
}
