% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{featureSelelction}
\alias{featureSelelction}
\title{featureSelelction}
\usage{
featureSelelction(sp.obj, sc.obj, n.features)
}
\arguments{
\item{sp.obj}{Seurat object of ST data.}

\item{sc.obj}{Seurat object of SC data.}

\item{n.features}{Common features between SC and ST data.}
}
\description{
featureSelelction
}
