% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMarkers.R
\name{findMarkersBySeurat}
\alias{findMarkersBySeurat}
\title{findMarkersBySeurat}
\usage{
findMarkersBySeurat(sc.obj, group.size, assay = "RNA", verbose = FALSE)
}
\arguments{
\item{sc.obj}{Seurat object of single-cell data.}

\item{group.size}{Marker size of each subset derived from SC data.}

\item{assay}{Assay type to use for SC data normalization. Default: RNA.}
}
\value{
A list of identified markers.
}
\description{
Find markers among cell types based on SC data using Seurat-inbuilt method.
}
