% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMarkers.R
\name{findScMarkersByShannon}
\alias{findScMarkersByShannon}
\title{findScMarkersByShannon}
\usage{
findScMarkersByShannon(
  sc.obj,
  group.size,
  assay = "RNA",
  min.cells.of.subset = 5,
  percent.cut = 0.25
)
}
\arguments{
\item{sc.obj}{Seurat object of single-cell data.}

\item{group.size}{Marker size of each subset derived from SC data.}

\item{assay}{Assay type to use for SC data normalization. Default: RNA.}

\item{min.cells.of.subset}{Include cells detected in at least one cell type in the SC data. Default: 5.}

\item{percent.cut}{Minimum fraction of cells in a cluster expressing a gene. Default: 0.25.}
}
\value{
A list of identified markers.
}
\description{
Find markers among cell types based on SC data using Shannon-entropy based strategy.
}
