% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoords.R
\name{getRandomCords}
\alias{getRandomCords}
\title{getRandomCords}
\usage{
getRandomCords(sp.coord, num.cells, seed = 123456, n.workers = 4)
}
\arguments{
\item{sp.coord}{Coordinates of ST image data.}

\item{num.cells}{A list of number of cells for mapping to spot.}

\item{seed}{Set seed for generating radom coordinates. Default: 123456.}

\item{n.workers}{Number of cores for parallel processing. Default: 4.}
}
\value{
A data.frame of generated spatial coordinates.
}
\description{
Get random coordinates for mapping single-cells to spatial.
}
