% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{hclustPartition}
\alias{hclustPartition}
\title{hclustPartition}
\usage{
hclustPartition(netx.pred, sp.obj, sc.obj, st.prop, num.cells)
}
\arguments{
\item{netx.pred}{A data.frame of probabilities inferred from FNN model.}

\item{sp.obj}{Seurat object containing ST data.}

\item{sc.obj}{Seurat object containing SC data.}

\item{st.prop}{Estimated proportions for ST spots.}

\item{num.cells}{A vector of cell counts in spots}
}
\value{
A list of single cells partitioned into sub-clusters.
}
\description{
hclustPartition
}
