% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightDist.R
\name{integDataBySeurat}
\alias{integDataBySeurat}
\title{integDataBySeurat}
\usage{
integDataBySeurat(
  sp.obj,
  sc.obj,
  assay,
  reduction,
  npcs = 30,
  n.features = 3000,
  dist.based = c("UMAP", "TSNE"),
  verbose = TRUE
)
}
\arguments{
\item{sp.obj}{Seurat object of spatial transcriptome (ST) data.}

\item{sc.obj}{Seurat object of single-cell data.}

\item{assay}{Assay type to use for SC data normalization.}

\item{reduction}{Dimensionality reduction technique to align spatial and single-cell data.}

\item{npcs}{Number of PCs for running UMAP. Default: 30.}

\item{n.features}{Number of features set for SelectIntegrationFeatures. Default: 3000.}

\item{dist.based}{Dimensionality reduction basis used for distance weighting, UMAP or TSNE. Default: UMAP.}

\item{verbose}{Show running messages or not. Default: TRUE.}
}
\value{
Integrated Seurat object.
}
\description{
Integrate SC and ST data normalized by the LogNormalize method using the CCA method.
}
