% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{linearSumAssignment}
\alias{linearSumAssignment}
\title{linearSumAssignment}
\usage{
linearSumAssignment(
  sp.obj,
  sc.obj,
  out.sim,
  num.cells,
  n.features,
  partion = TRUE
)
}
\arguments{
\item{sp.obj}{Seurat object of ST data.}

\item{sc.obj}{Seurat object of SC data.}

\item{out.sim}{Similarity matrix between single-cells and spots.}

\item{num.cells}{A vector of cell count per spot projected to spots. Default: 10.}

\item{n.features}{Number of common features between SC and ST data.}

\item{partion}{Split into sub-modules mapped to spatial positions. Default: TRUE.}
}
\value{
A list of assigned cells corresponding to spots.
}
\description{
Assign single-cells to spatial coordinates.
}
