% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{mergeClusters}
\alias{mergeClusters}
\title{mergeClusters}
\usage{
mergeClusters(sp.obj, num.cells, max.block.size = 20000)
}
\arguments{
\item{sp.obj}{Seurat object containing single-cell or spatial transcriptomics data.}

\item{num.cells}{A vector of cell counts in spots.}

\item{max.block.size}{Numeric, the maximum size of the block for dividing the clusters. Default is 20000.}
}
\value{
A vector of merged cluster identifiers.
}
\description{
Merge clusters in a Seurat object based on a maximum block size, integrating single-cell or spatial transcriptomics data.
}
