% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multipleProcess}
\alias{multipleProcess}
\title{multipleProcess}
\usage{
multipleProcess(n.workers = 4)
}
\arguments{
\item{n.workers}{Number of cores. Default: 4.}
}
\description{
Open multiple workers for Seurat processing.
}
