% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{partionClusters}
\alias{partionClusters}
\title{partionClusters}
\usage{
partionClusters(sp.obj, sc.obj, num.cells, hot.spts)
}
\arguments{
\item{sp.obj}{Seurat object containing ST data.}

\item{sc.obj}{Seurat object containing SC data.}

\item{num.cells}{A vector of cell counts in spots.}
}
\value{
A list of single cells partitioned into sub-clusters.
}
\description{
Partition single cells into sub-clusters using a deep learning strategy based on integration of spatial transcriptomics (ST) and single-cell (SC) data.
}
