% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{partitionClusters}
\alias{partitionClusters}
\title{partitionClusters}
\usage{
partitionClusters(
  sp.obj,
  sc.obj,
  hot.pvals,
  st.prop.cnts,
  assay = "SCT",
  reduction = "cca",
  npcs = 30
)
}
\arguments{
\item{sp.obj}{Seurat object containing ST data.}

\item{sc.obj}{Seurat object containing SC data.}

\item{hot.pvals}{Data frame of hotspot p-values indicating hotspot presence in spots and cell types.}

\item{st.prop.cnts}{Estimated spot-level counts matrix (rows=spots, cols=cell types).}

\item{assay}{Assay to use for SC data. Default: "SCT".}

\item{reduction}{Dimensionality reduction technique to align spatial and single-cell data. Default: cca.}

\item{npcs}{Number of PCA components to use. Default: 30.}
}
\value{
A list of single cells partitioned into sub-clusters.
}
\description{
Partition single cells into sub-clusters using a deep learning strategy based on integration of spatial transcriptomics (ST) and single-cell (SC) data.
}
