% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{plotCoexistCtypes}
\alias{plotCoexistCtypes}
\title{plotCoexistCtypes}
\usage{
plotCoexistCtypes(
  j.index,
  j.cut = 0,
  high.ctypes = NULL,
  cols = c("#91a28c", "white", "#8f2c37")
)
}
\arguments{
\item{j.index}{A data.frame of coexist indexes of cell-types.}

\item{j.cut}{J-index value below than the specified cutoff will be set to zero. Default: 0.}

\item{cols}{A vector colors with names, which contains three colors, representing low, medium, and high. Default: c("#91a28c", "white", "#8f2c37").}
}
\value{
ggplot object.
}
\description{
Showing coexist indexes of cell-types.
}
