% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{plotSc2Spatial}
\alias{plotSc2Spatial}
\title{plotSc2Spatial}
\usage{
plotSc2Spatial(
  sp.obj,
  sce,
  tar.ctypes = NULL,
  pt.size = 1,
  colors = NULL,
  shape = 21,
  percent = NULL
)
}
\arguments{
\item{sp.obj}{Seurat object of spatial transcriptome (ST) data.}

\item{sce}{SingleCellExperiment object represents the output generated by the Cell2Spatial tool after projecting single-cell data onto spatial coordinates}

\item{tar.ctypes}{A vector of cell types to show exclusively. Default: NULL.}

\item{pt.size}{Point size of single-cells on spatial image. Default: 1.0.}

\item{colors}{A vector colors corresponding to the cell type. Default: NULL.}

\item{shape}{Point shape. Default: 21.}

\item{percent}{Down-sampling cell types if percent is not NULL (0 < percent <= 1).}
}
\value{
ggplot object
}
\description{
Visualizing the distribution of single-cells on spatial coordinates based on SingleCellExperiment object.
}
