% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocessSeqData}
\alias{preprocessSeqData}
\title{preprocessSeqData}
\usage{
preprocessSeqData(
  sp.obj,
  sc.obj,
  cell.type.column,
  normalize.method,
  min.cells.of.subset = 5,
  verbose = TRUE
)
}
\arguments{
\item{sp.obj}{Seurat object containing spatial transcriptomics data.}

\item{sc.obj}{Seurat object containing single-cell RNA-seq data.}

\item{cell.type.column}{Column name in the meta.data slot of SC Seurat object specifying cell types.}

\item{normalize.method}{Normalization method for SC and ST data, "SCTransform" or "LogNormalize".}

\item{min.cells.of.subset}{Minimum number of cells required per cell type in the SC data. Default: 5.}

\item{verbose}{Boolean indicating whether to show progress messages. Default: TRUE.}
}
\value{
A list containing preprocessed SC and ST Seurat objects.
}
\description{
Preprocesses single-cell (SC) and spatial transcriptomics (ST) data by down-sampling, filtering low-quality cell types, and normalizing both datasets.
}
