% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{println}
\alias{println}
\title{println}
\usage{
<<<<<<< HEAD
println(X, verbose = TRUE, status = c("INFO", "ERROR", "WARN"), ...)
=======
println(X, verbose = TRUE, status = c("INFO", "ERROR"), ...)
>>>>>>> 5859e3527b186cec1e7cc9974128adabbcd95a88
}
\arguments{
\item{verbose}{Bool value, print out message or not, default: TRUE.}

<<<<<<< HEAD
\item{status}{Normal running messages, warn, or error. Default: INFO (normal status).}
=======
\item{status}{Normal running messages or error. Default: INFO (normal status).}
>>>>>>> 5859e3527b186cec1e7cc9974128adabbcd95a88

\item{infos}{Message that need to be printed.}
}
\value{
NULL.
}
\description{
Print out message if verbose equals TRUE.
}
