% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prop2counts}
\alias{prop2counts}
\title{prop2counts}
\usage{
prop2counts(st.prop, num.cells)
}
\arguments{
\item{st.prop}{A matrix of proportions, where each row represents a different spot (e.g., tissue region or location), and each column represents a different cell type or category.}

\item{num.cells}{A vector of length equal to the number of rows in \code{st.prop}, representing the total number of cells available for each corresponding spot.}
}
\value{
A matrix of integer cell counts corresponding to the proportions in \code{st.prop}, ensuring the total number of cells for each spot matches \code{num.cells}.
}
\description{
Converts proportions to integer cell counts based on a given number of cells for each spot.
}
