% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightDist.R
\name{psedoSpotExprUseSC}
\alias{psedoSpotExprUseSC}
\title{psedoSpotExprUseSC}
\usage{
psedoSpotExprUseSC(sc.obj, sp.obj, pseu.cnt = 200, lamba = 5, mc.cores = 4)
}
\arguments{
\item{sc.obj}{Seurat object of SC data.}

\item{sp.obj}{Seurat object of ST data.}

\item{pseu.cnt}{Pseudo cell counts for each cell type from SC data. Default: 200.}

\item{lamba}{Median of cell counts for spots.}

\item{mc.cores}{Number of cores for parallel running. Default: 4}
}
\value{
Seurat object of pseudo purity spots data.
}
\description{
Pseudo purity spots expression based on SC data.
}
