% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightDist.R
\name{pseudoSpotExprUseSC}
\alias{pseudoSpotExprUseSC}
\title{pseudoSpotExprUseSC}
\usage{
pseudoSpotExprUseSC(sc.obj, sp.obj, assay, pseu.cnt = 200, lamba = 5)
}
\arguments{
\item{sc.obj}{Seurat object of SC data.}

\item{sp.obj}{Seurat object of ST data.}

\item{assay}{Assay type to use for SC data normalization.}

\item{pseu.cnt}{Pseudo cell counts for each cell type from SC data. Default: 200.}

\item{lamba}{Median of cell counts for spots.}
}
\value{
Seurat object of pseudo purity spots data.
}
\description{
Pseudo purity spots expression based on SC data.
}
