% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMarkers.R
\name{qualityGenesIndex}
\alias{qualityGenesIndex}
\title{qualityGenesIndex}
\usage{
qualityGenesIndex(sc.obj, assay = "RNA")
}
\arguments{
\item{sc.obj}{Seurat object of single-cell data.}

\item{assay}{Assay type to use for SC data normalization. Default: RNA.}
}
\value{
A list of indexes for genes among cell types.
}
\description{
Calculation of quality indexes for genes among cell types.
}
