% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMap2Sp.R
\name{runMap2SP}
\alias{runMap2SP}
\title{runMap2SP}
\usage{
runMap2SP(
  sp.obj,
  sc.obj,
  sc.markers = NULL,
  ctype = "idents",
  group.size = 30,
  res = 0.8,
  integ = TRUE,
  duplicated = FALSE,
  partion = TRUE,
  min.cells.of.subset = 5,
  max.cells.in.spot = 10,
  fix.cells.in.spot = NULL,
  n.features = 3000,
  knn = 5,
  sample.size = NULL,
  select.markers = c("shannon", "wilcox"),
  dist.method = c("mle", "cor"),
  return.type = c("Seurat", "SingleCellExperiment"),
  n.workers = 4,
  verbose = TRUE
)
}
\arguments{
\item{sp.obj}{Seurat object of spatial transcriptome (ST) data.}

\item{sc.obj}{Seurat object of single-cell (SC) data.}

\item{sc.markers}{A list of markers for cell types. If not provided, automated inference is performed using the modified Shannon-entropy method. Default: NULL.}

\item{ctype}{Specify the column name for the cell type in meta.data slot of the SC Seurat object. Default: idents.}

\item{group.size}{Specify the marker size for each subset derived from single-cell data. This information is crucial for estimating the activity of markers within each cell type. Default: 30.}

\item{res}{Resolution for clustering ST spot. Default: 0.8.}

\item{integ}{Integration of SC and ST data or not. For high-resolution ST data, please set FALSE to accelerate the running time. Default: TRUE.}

\item{duplicated}{Assigning individual cells to ST coordinates with duplicated or not. Default: FALSE.}

\item{partion}{Split into sub-modules mapped to spatial positions when 'duplicated' set to TRUE. Default: TRUE.}

\item{min.cells.of.subset}{Include cells detected in at least one cell type in the SC data. Default: 5.}

\item{max.cells.in.spot}{Maximum number of cells in ST spots. Default: 10 (for 10x Visium).}

\item{fix.cells.in.spot}{Fixed number of cells assigned to a spot. Default: NULL.}

\item{n.features}{Top k highly variable features used for integrating SC and ST data. Default: 3000.}

\item{knn}{Utilizing the k nearest spots to filter out spots that do not contain cell types present in the SC reference. Default: 5.}

\item{sample.size}{Down-sampling a small subset of SC data for mapping to ST coordinates. Default: NULL.}

\item{select.markers}{Method for selecting cell-type specific markers when 'sc.markers = NULL': modified Shannon-entropy strategy (shannon) or Wilcoxon test (wilcox) implemented by the Seurat package. Default: shannon.}

\item{dist.method}{Measure the distance between single-cell and spots, maximum likehood model (mle) or correlation (cor). Default: mle.}

\item{return.type}{Assigned results can be of the object type Seurat or SingleCellExperiment. Default: Seurat.}

\item{verbose}{Show running messages or not. Default: TRUE.}

\item{n.wrokers}{Number of cores for parallel processing. Default: 4.}
}
\value{
Assigning results wrapped as a Seurat or SingleCellExperiment object.
}
\description{
Accurate mapping of single cells to spatial coordinates.
}
\examples{
sp.obj <- system.file("data", "Kindney_SP.RDS", package = "Cell2Spatial") \%>\% readRDS(.)
sc.obj <- system.file("data", "Kindney_SC.RDS", package = "Cell2Spatial") \%>\% readRDS(.)
sce <- runMap2SP(sp.obj, sc.obj, ctype = "mainCtype", res = 0.8, group.size = 30)
}
