% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sampleBulkProfiles}
\alias{sampleBulkProfiles}
\title{sampleBulkProfiles}
\usage{
sampleBulkProfiles(sp.obj, base.ref, nperm = 100)
}
\arguments{
\item{base.ref}{Base matrix.}

\item{nperm}{Number of permutations. Default: 100.}
}
\value{
A list of generated bulk profiles.
}
\description{
Down-sampling bulk profiles for deconvoluting.
}
