% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{selectByProb}
\alias{selectByProb}
\title{selectByProb}
\usage{
selectByProb(sp.score, sc.score)
}
\arguments{
\item{sp.score}{Matrix or data frame of ST signature scores.}

\item{sc.score}{Matrix or data frame of SC signature scores.}
}
\value{
Similarity matrix where each row represents a spatial spot and each column represents a single-cell, adjusted based on computed probabilities.
}
\description{
Computes a similarity score matrix based on spatial and single-cell signature scores, adjusting for probabilities.
}
