% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkParams.R
\name{selectMakers}
\alias{selectMakers}
\title{selectMakers}
\usage{
selectMakers(sc.obj, sc.markers, select.markers, group.size, verbose = TRUE)
}
\arguments{
\item{sc.obj}{Seurat object of SC data.}

\item{sc.markers}{When the number of cell types identified from SC data is one, markers specific to that cell type must be provided. Default: NULL.}

\item{select.markers}{Method for selecting cell-type specific markers when \code{sc.markers = NULL}.}

\item{group.size}{Specify the marker size for each subset derived from single-cell data.}

\item{verbose}{Boolean indicating whether to show running messages. Default: TRUE.}
}
\value{
A list of markers for cell types.
}
\description{
Select marker genes for each cell type based on single-cell (SC) data.
}
