% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{selectScByCor}
\alias{selectScByCor}
\title{selectScByCor}
\usage{
selectScByCor(sp.score, sc.score, adj.w)
}
\arguments{
\item{sp.score}{Score matrix of ST inferred from a list of markers.}

\item{sc.score}{Score matrix of SC inferred from a list of markers.}

\item{adj.w}{A weighted matrix infered from distance. Rows represent spot names, and cloumns are single-cells.}
}
\value{
Similarity matrix.
}
\description{
Select single-cells based on weighted correlation.
}
