% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightDist.R
\name{weightDist}
\alias{weightDist}
\title{weightDist}
\usage{
weightDist(
  sc.obj,
  sp.obj,
  lamba,
  assay,
  reduction,
  cells.thresh = 50000,
  dist.based = c("UMAP", "TSNE")
)
}
\arguments{
\item{sc.obj}{Seurat object of SC data.}

\item{sp.obj}{Seurat object of ST data.}

\item{lamba}{Median of cell counts for spots.}

\item{assay}{The assay to use for the normalize analysis.}

\item{reduction}{Dimensionality reduction technique to align spatial and single-cell data.}

\item{cells.thresh}{Threshold for cell counts in the SC object. When the cell count exceeds this threshold, the function switches to a faster mode. Default: 50000.}

\item{dist.based}{Dimensionality reduction basis used for distance weighting. Default: UMAP.}
}
\value{
A matrix of weighted distance matrix. Rows represent spot clusters and columns are cell types.
}
\description{
Weight distance between SC and ST data.
}
