% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{weightSimScore}
\alias{weightSimScore}
\title{weightSimScore}
\usage{
weightSimScore(out.sim, adj.w, spot.name, cell.names, hot.spts)
}
\arguments{
\item{out.sim}{A subset of similariy scores between SC and ST.}

\item{adj.w}{A weighted matrix infered from distance. Rows represent spot names, and cloumns are cell types.}

\item{spot.name}{Target spot names.}

\item{cell.names}{Target single-cell names.}

\item{hot.spts}{A Boolean value matrix indicates whether the hot spots belong to a specific cell type or not.}
}
\value{
Weighted similarity matrix.
}
\description{
Weighted similarity score based on spatial distances.
}
