% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{weigthNetProb}
\alias{weigthNetProb}
\title{weigthNetProb}
\usage{
weigthNetProb(netx.pred, sc.obj, sp.obj, hot.spts)
}
\arguments{
\item{netx.pred}{Matrix or data frame of network predictions.}

\item{sc.obj}{Seurat object of SC data.}

\item{sp.obj}{Seurat object of ST data.}

\item{hot.spts}{A data frame indicating hotspot presence in spots and cell types.}
}
\value{
Adjusted network predictions matrix where weights are applied based on cell type proportions and hotspot information.
}
\description{
Adjusts FNN predictions based on cell type proportions derived from SC and ST data, considering hotspot information.
}
