% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightDist.R
\name{adjcentScOfSPGlobal}
\alias{adjcentScOfSPGlobal}
\title{adjcentScOfSPGlobal}
\usage{
adjcentScOfSPGlobal(obj, quantile.cut = 0.3)
}
\arguments{
\item{obj}{Integrated Seurat object.}

\item{quantile.cut}{Numeric value specifying the quantile threshold for distance scaling. Default is 0.30, which considers the maximum distance for normalization.}
}
\value{
A matrix of distance weights between ST and SC.
}
\description{
Generating distance matrix between ST spots and SC cells.
}
