% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estCellPerSpots.R
\name{adjustScObj}
\alias{adjustScObj}
\title{adjustScObj}
\usage{
adjustScObj(sc.obj, st.prop, num.cells)
}
\arguments{
\item{sc.obj}{Seurat object of single-cell (SC) data.}

\item{st.prop}{A data.frame containing proportions for ST data.}

\item{num.cells}{A vector of cell counts for ST spots.}
}
\value{
Adjusted Seurat object of SC data.
}
\description{
Adjust single-cell (SC) data by incorporating inferred proportions and cell counts for mapping to spatial transcriptomics (ST) spots
}
