% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{assignSc2SP}
\alias{assignSc2SP}
\title{assignSc2SP}
\usage{
<<<<<<< HEAD
assignSc2SP(
  sp.obj,
  sc.obj,
  out.sim,
  num.cells,
  st.pvals,
  n.features,
  duplicated = FALSE,
  partion = TRUE,
  p.cut = 0.05
)
=======
assignSc2SP(sp.obj, sc.obj, out.sim, num.cells, st.pvals, duplicated = FALSE)
>>>>>>> 5859e3527b186cec1e7cc9974128adabbcd95a88
}
\arguments{
\item{sp.obj}{Seurat object of ST data.}

\item{sc.obj}{Seurat object of SC data.}

\item{out.sim}{Similarity matrix between single-cells and spots.}

<<<<<<< HEAD
\item{num.cells}{A vector of cell count per spot projected to spots. Default: 10.}

\item{st.pvals}{A data.frame of p-values of ST spots.}

\item{n.features}{Number of common features between SC and ST data.}

\item{duplicated}{Unique assign single-cells to spot or not. Default: FALSE.}

\item{p.cut}{Cutoff pvalue to determine negative spots. Default: 0.05}

\item{Split}{into sub-modules mapped to spatial positions. Default: TRUE.}
=======
\item{num.cells}{Number of cells projected to spots. Default: 10.}

\item{st.pvals}{A data.frame of p-values of ST spots.}

\item{duplicated}{Unique assign single-cells to spot or not. Default: FALSE.}
>>>>>>> 5859e3527b186cec1e7cc9974128adabbcd95a88
}
\value{
A list of assigned cells corresponding to spots.
}
\description{
Assign single-cells to spots using linear assignment method.
}
