% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoords.R
\name{assignSCcords}
\alias{assignSCcords}
\title{assignSCcords}
\usage{
assignSCcords(sp.obj, sc.obj, out.sc, num.cells)
}
\arguments{
\item{sp.obj}{Seurat object of spatial transcriptome (ST) data.}

\item{sc.obj}{Seurat object of single-cell data.}

\item{out.sc}{A list of ordered single-cells correspond to spots.}

\item{num.cells}{A list of number of cells for mapping to spot.}
}
\value{
Seurat object for mapped results.
}
\description{
Assign single-cells to spatial coordinates.
}
