% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoords.R
\name{calcSpotsDist}
\alias{calcSpotsDist}
\title{calcSpotsDist}
\usage{
<<<<<<< HEAD
calcSpotsDist(sp.obj, knn = 5)
=======
calcSpotsDist(sp.obj)
>>>>>>> 5859e3527b186cec1e7cc9974128adabbcd95a88
}
\arguments{
\item{sp.obj}{Spatial seurat object data.}
}
\value{
<<<<<<< HEAD
A list distance relationship.
=======
Eulerian distance matrix.
>>>>>>> 5859e3527b186cec1e7cc9974128adabbcd95a88
}
\description{
Identified significant spots located in medulla areas and calculate the distances between spots.
}
