% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoords.R
\name{calcSpotsDist}
\alias{calcSpotsDist}
\title{calcSpotsDist}
\usage{
calcSpotsDist(sp.obj, knn = 5)
}
\arguments{
\item{sp.obj}{Spatial seurat object data.}
}
\value{
A list distance relationship.
}
\description{
Identified significant spots located in medulla areas and calculate the distances between spots.
}
