% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{ctypesOfClusters}
\alias{ctypesOfClusters}
\title{ctypesOfClusters}
\usage{
ctypesOfClusters(sp.obj, hot.spts)
}
\arguments{
\item{sp.obj}{Seurat object of ST data.}

\item{hot.spts}{A data.frame containing boolean values to determine whether spots are hotspots or not.}
}
\value{
A data frame where each row represents a cluster and each column represents a cellular type, with values indicating weighted scores.
}
\description{
Calculates weighted scores of cellular types across clusters based on spatial transcriptomics (ST) data and hotspot information.
}
