% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{ctypesOfClusters}
\alias{ctypesOfClusters}
\title{ctypesOfClusters}
\usage{
ctypesOfClusters(sp.obj, hot.pvals, p.cut = 0.05)
}
\arguments{
\item{sp.obj}{Seurat object of ST data.}

\item{hot.pvals}{A data.frame containing p-values generated by hotspot detection step.}

\item{p.cut}{Threashold value to filter out low-condfi}
}
\value{
A data frame where each row represents a cluster and each column represents a cellular type, with values indicating weighted scores.
}
\description{
Calculates weighted scores of cellular types across clusters based on spatial transcriptomics (ST) data and hotspot information.
}
