% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estCellPerSpots.R
\name{estCellPerSpots}
\alias{estCellPerSpots}
\title{estCellPerSpots}
\usage{
<<<<<<< HEAD
estCellPerSpots(
  sp.obj,
  max.cells = 10,
  fix.cells.in.spot = NULL,
  quantile.cut = 0.99,
  num.genes = 2000
)
=======
estCellPerSpots(sp.obj, max.cells = 10, quantile.cut = 0.99, num.genes = 2000)
>>>>>>> 5859e3527b186cec1e7cc9974128adabbcd95a88
}
\arguments{
\item{sp.obj}{Seurat object of spatial transcriptome (ST) data.}

\item{max.cells}{Maximum number of cells in Spots. Default: 10 (10x Genomics).}
<<<<<<< HEAD

\item{fix.cells.in.spot}{Fixed number of cells assigned to a spot. Default: NULL.}

\item{quantile.cut}{Spots with a number of UMIs greater than the quantile cutoff are set as a reference. Default: 0.99.}

\item{num.genes}{Set the number of stable genes. Default: 2000.}
=======
>>>>>>> 5859e3527b186cec1e7cc9974128adabbcd95a88
}
\value{
A vector of estimated cells per spot.
}
\description{
Estimate number of cells per spot.
}
