% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{estPropInSpots}
\alias{estPropInSpots}
\title{estPropInSpots}
\usage{
estPropInSpots(
  sp.obj,
  sc.obj,
  st.pvals,
  sc.markers,
  pval.cut = 0.05,
  knn = 5,
<<<<<<< HEAD
  intercept = TRUE,
  chunk.size = 10000
=======
  intercept = TRUE
>>>>>>> 5859e3527b186cec1e7cc9974128adabbcd95a88
)
}
\arguments{
\item{sp.obj}{Seurat object of spatial transcriptome (ST) data.}

\item{sc.obj}{Seurat object of single-cell (SC) data.}

\item{st.pvals}{A data.frame containing p-values to determine whether cell types are significantly present in a spot or not.}

\item{sc.markers}{A list of markers of cell types.}

\item{pval.cut}{significant cutoff value. Default: 0.05}

\item{knn}{K nearest cells used for estimating cellular proportions. Default: 5.}

\item{intercept}{Using intercept or not for robust linear regression. Default: TRUE.}
}
\value{
A data frame of cellular fractions in each spot.
}
\description{
Infer whether a cell type belongs to a specific cluster.
}
