% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{estPropInSpots}
\alias{estPropInSpots}
\title{estPropInSpots}
\usage{
estPropInSpots(
  sp.obj,
  sc.obj,
  hot.pvals,
  sc.markers,
  knn = 5,
  intercept = TRUE,
  chunk.size = 10000
)
}
\arguments{
\item{sp.obj}{Seurat object of ST data.}

\item{sc.obj}{Seurat object of single-cell (SC) data.}

\item{hot.pvals}{A data.frame containing p-values generated by hotspot detection step.}

\item{sc.markers}{A list of markers of cell types.}

\item{knn}{K nearest cells used for estimating cellular proportions. Default: 5.}

\item{intercept}{Boolean indicating whether to use intercept for robust linear regression. Default: TRUE.}

\item{chunk.size}{Chunk size to determine downsampled ST data for estimating proportions. Default: 10000.}
}
\value{
A data.frame of estimated cellular proportions in spots.
}
\description{
Infer cellular proportions for spatial transcriptome (ST) spots.
}
