% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv.R
\name{estPropInSpots}
\alias{estPropInSpots}
\title{estPropInSpots}
\usage{
estPropInSpots(
  sp.obj,
  sc.obj,
  sc.markers,
  num.cells = NULL,
  hot.pvals = NULL,
  weight = TRUE,
  assay = "SCT",
  knn = 5,
  intercept = TRUE
)
}
\arguments{
\item{sp.obj}{Seurat object of ST data.}

\item{sc.obj}{Seurat object of single-cell (SC) data.}

\item{sc.markers}{A list of markers of cell types.}

\item{hot.pvals}{A data.frame containing p-values generated by hotspot detection step. Default: NULL.}

\item{weight}{Logical, indicating whether to incorporate hotspot weights in the estimation. Default: TRUE.}

\item{assay}{The assay to use for SC data. Default: "SCT" (Seurat's SCTransform).}

\item{knn}{K nearest cells used for estimating cellular proportions. Default: 5.}

\item{intercept}{Boolean indicating whether to use intercept for robust linear regression. Default: TRUE.}
}
\value{
A data.frame of estimated cellular proportions in spots.
}
\description{
Infer cellular proportions for spatial transcriptome (ST) spots.
}
