% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{featureSelelction}
\alias{featureSelelction}
\title{featureSelelction}
\usage{
featureSelelction(
  sp.obj,
  sc.obj,
  n.features = 3000,
  assay = "RNA",
  verbose = TRUE
)
}
\arguments{
\item{sp.obj}{Seurat object containing ST data.}

\item{sc.obj}{Seurat object containing SC data.}

\item{n.features}{Number of features to select for integration. Default: 3000.}

\item{verbose}{Boolean indicating whether to display verbose messages during processing. Default: TRUE.}
}
\value{
A list containing integrated Seurat objects for SC (\code{sc.int}) and ST (\code{st.int}) data.
}
\description{
Performs feature selection and data integration between spatial transcriptomics (ST) and single-cell (SC) data.
}
