% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightDist.R
\name{findClustersForSpData}
\alias{findClustersForSpData}
\title{findClustersForSpData}
\usage{
findClustersForSpData(obj.seu, npcs = 30, res = 0.8, verbose = TRUE)
}
\arguments{
\item{obj.seu}{Seurat object of ST data.}

\item{npcs}{Number of PCs for clustering. Default: 50.}

\item{res}{Resolution parameter for finding clusters. Default: 0.8.}
}
\value{
Seurat object after clustering.
}
\description{
Find clusters for single-cell data based on Seurat object.
}
