% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectHotSpots.R
\name{getGsetScore}
\alias{getGsetScore}
\title{getGsetScore}
\usage{
getGsetScore(obj.seu, sc.markers, assay = "SCT")
}
\arguments{
\item{obj.seu}{Seurat object.}

\item{sc.markers}{A list of markers for cell types.}

\item{assay}{The assay used to calculate signature scores of cell types. Default: SCT.}
}
\value{
A data.frame of signature scores.
}
\description{
Get signature scores in each cell or spot using a list of markers.
}
