% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estCellPerSpots.R
\name{inferCellNumbers}
\alias{inferCellNumbers}
\title{inferCellNumbers}
\usage{
inferCellNumbers(
  sp.obj,
  max.cells = 10,
  umi.thresh = 2,
  w.L = 0.5,
  w.T = 0.5,
  fix.cells.in.spot = NULL
)
}
\arguments{
\item{sp.obj}{A Seurat object containing spatial transcriptomics data (Q matrix in the "counts" slot of the "Spatial" assay).}

\item{max.cells}{Maximum number of cells assumed in any spot (default = 10).}

\item{umi.thresh}{Minimum UMI count to consider a gene as expressed (default = 2).}

\item{w.L}{Weight for library size in the composite score (default = 0.5).}

\item{w.T}{Weight for gene count in the composite score (default = 0.5).}

\item{fix.cells.in.spot}{Fixed number of cells assigned to each spot. Default: NULL.}
}
\value{
A named vector of estimated cell numbers per spot.
}
\description{
Estimates the number of cells per spot in a spatial transcriptomics Seurat object using a corrected saturation model.
}
