% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{integDataBySeurat}
\alias{integDataBySeurat}
\title{integDataBySeurat}
\usage{
integDataBySeurat(
  sp.obj,
  sc.obj,
  npcs = 30,
  n.features = 3000,
  integ = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{sp.obj}{Seurat object of spatial transcriptome (ST) data.}

\item{sc.obj}{Seurat object of single-cell data.}

\item{npcs}{Number of PCs for running UMAP. Default: 30.}

\item{integ}{Integrate SC and ST data or not. Default: TRUE.}

\item{verbose}{Show running messages or not. Default: TRUE.}
}
\value{
A list of information about integration.
}
\description{
Integrate SC and ST data using Harmony method.
}
