% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{linearSumAssignment}
\alias{linearSumAssignment}
\title{linearSumAssignment}
\usage{
linearSumAssignment(
  sp.obj,
  sc.obj,
  out.sim,
  adj.w,
  num.cells,
  hot.spts,
  partition,
  hclust = TRUE
)
}
\arguments{
\item{sp.obj}{Seurat object containing ST data.}

\item{sc.obj}{Seurat object containing SC data.}

\item{out.sim}{Similarity matrix between single cells and spots.}

\item{adj.w}{Weight matrix for adjusting similarity scores.}

\item{num.cells}{A vector indicating the cell count per spot projected to spots.}

\item{hot.spts}{Data frame indicating hotspot presence in spots and cell types.}

\item{partition}{Logical indicating whether to split into sub-modules mapped to spatial positions.}

\item{hclust}{Logical. If TRUE, hierarchical clustering is applied to group ST clusters. Default: TRUE.}
}
\value{
A list of assigned cells corresponding to spots.
}
\description{
Assigns single cells to spatial coordinates based on similarity scores between single-cell and spatial transcriptomics (ST) data.
}
