% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{linearSumAssignment}
\alias{linearSumAssignment}
\title{linearSumAssignment}
\usage{
linearSumAssignment(
  sp.obj,
  sc.obj,
  out.sim,
  adj.w,
  hot.pvals,
  st.prop.cnts,
  num.cells,
  partition,
  assay,
  reduction,
  thresh.cells = 30000
)
}
\arguments{
\item{sp.obj}{Seurat object containing ST data.}

\item{sc.obj}{Seurat object containing SC data.}

\item{out.sim}{Similarity matrix between single cells and spots.}

\item{adj.w}{Weight matrix for adjusting similarity scores.}

\item{hot.pvals}{Data frame of hotspot p-values indicating hotspot presence in spots and cell types.}

\item{st.prop.cnts}{Estimated spot-level counts matrix (rows=spots, cols=cell types).}

\item{num.cells}{A vector of cell counts for each spot.}

\item{partition}{Logical indicating whether to split into sub-modules mapped to spatial positions.}

\item{assay}{Assay to use for SC data. Default: "SCT".}

\item{reduction}{Dimensionality reduction technique to align spatial and single-cell data. Default: cca.}

\item{thresh.cells}{If the number of cells exceeds this threshold, the partitioning strategy is executed automatically. Default: 30000.}
}
\value{
A list of assigned cells corresponding to spots.
}
\description{
Assigns single cells to spatial coordinates based on similarity scores between single-cell and spatial transcriptomics (ST) data.
}
