% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{partionClusters}
\alias{partionClusters}
\title{partionClusters}
\usage{
partionClusters(sp.obj, sc.obj, n.features, num.cells)
}
\arguments{
\item{sp.obj}{Seurat object of ST data.}

\item{sc.obj}{Seurat object of SC data.}

\item{n.features}{Number of common features between SC and ST data.}

\item{num.cells}{A vector of cell counts in spots.}
}
\value{
A list of single-cells in each cluster.
}
\description{
Partion single-cells to sub-clusters using deep learning strategy.
}
