% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{partitionClusters}
\alias{partitionClusters}
\title{partitionClusters}
\usage{
partitionClusters(sp.obj, sc.obj, num.cells, hot.spts, st.prop, hclust = TRUE)
}
\arguments{
\item{sp.obj}{Seurat object containing ST data.}

\item{sc.obj}{Seurat object containing SC data.}

\item{num.cells}{A vector of cell counts in spots.}

\item{hot.spts}{Data frame indicating hotspot presence in spots and cell types.}

\item{st.prop}{Estimated proportions for ST spots.}

\item{hclust}{Logical. If TRUE, hierarchical clustering is applied to group ST clusters. Default: TRUE.}
}
\value{
A list of single cells partitioned into sub-clusters.
}
\description{
Partition single cells into sub-clusters using a deep learning strategy based on integration of spatial transcriptomics (ST) and single-cell (SC) data.
}
