% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkParams.R
\name{preprocessSeqData}
\alias{preprocessSeqData}
\title{preprocessSeqData}
\usage{
preprocessSeqData(
  sp.obj,
  sc.obj,
  sample.size,
  cell.type.column,
  min.cells.of.subset = 5,
  verbose = TRUE
)
}
\arguments{
\item{sp.obj}{Seurat object of ST data.}

\item{sc.obj}{Seurat object of SC data.}

\item{sample.size}{Proportion for downsampling single cells of each cell type.}

\item{cell.type.column}{Specify the column name for the cell type in the meta.data slot of the SC Seurat object. Default: idents.}

\item{min.cells.of.subset}{Include cells detected in at least one cell type in the SC data. Default: 5.}

\item{verbose}{Boolean indicating whether to show running messages. Default: TRUE.}
}
\value{
A list of Seurat objects of preprocessed data.
}
\description{
The preprocessing of single-cell (SC) data includes several steps such as down-sampling, filtering out low-quality cell types, and normalization.
}
