% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkParams.R
\name{preprocessSeqData}
\alias{preprocessSeqData}
\title{preprocessSeqData}
\usage{
preprocessSeqData(
  sp.obj,
  sc.obj,
  sample.size,
  ctype,
  min.cells.of.subset = 5,
  verbose = TRUE
)
}
\arguments{
\item{sp.obj}{Seurat object of ST data.}

\item{sc.obj}{Seurat object of SC data.}

\item{sample.size}{Down-sampling size for each subset of SC data.}

\item{ctype}{Ident information of SC data in the column name of meta.data slot.}

\item{min.cells.of.subset}{Minimum cell counts of each subset source from SC data. Default: 5.}

\item{verbose}{Showing running mesaeg or not. Default: TRUE.}
}
\value{
Seurat object of preprocessed single cell data.
}
\description{
Preprocessing of single-cell (SC) data involves a series of steps.
}
