% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{println}
\alias{println}
\title{println}
\usage{
println(X, ..., verbose = TRUE, status = c("INFO", "WARN", "ERROR"))
}
\arguments{
\item{verbose}{Bool value, print out message or not. Default: TRUE.}

\item{status}{Normal running messages, warn, or error. Default: INFO (normal status).}

\item{infos}{Message that need to be printed.}
}
\value{
NULL.
}
\description{
Print out message if verbose equals TRUE.
}
