% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMap2Sp.R
\name{runCell2Spatial}
\alias{runCell2Spatial}
\title{runCell2Spatial}
\usage{
runCell2Spatial(
  sp.obj,
  sc.obj,
  cell.type.column = NULL,
  cell.type.markers = NULL,
  normalize.method = c("SCTransform", "LogNormalize"),
  marker.selection = c("shannon", "wilcox", "logFC"),
  group.size = 30,
  knn.spots = 5,
  resolution = 0.8,
  max.cells.in.spot = 10,
  fix.cells.in.spot = FALSE,
  signature.scoring.method = c("AddModuleScore", "UCell", "AverageExpr"),
  hotspot.detection.threshold = 1,
  adjust.deconv = c("Scales", "NONE"),
  feature.based = c("gene.based", "celltype.based"),
  dist.based = c("UMAP", "TSNE"),
  spatial.weight = FALSE,
  reduction = c("cca", "rpca"),
  platform.res = c("Low", "High"),
  partition = FALSE,
  n.workers = 4,
  verbose = TRUE
)
}
\arguments{
\item{sp.obj}{Seurat object of spatial transcriptome (ST) data.}

\item{sc.obj}{Seurat object of single-cell (SC) data.}

\item{cell.type.column}{Specify the column name for the cell type in the meta.data slot of the SC Seurat object. Default: NULL.}

\item{cell.type.markers}{When the number of cell types identified from SC data is one, markers specific to that cell type must be provided. Default: NULL.}

\item{normalize.method}{Normalization method for scRNA-seq data. Default: SCTransform.}

\item{marker.selection}{Method for selecting cell-type specific markers when \code{sc.markers = NULL}: modified Shannon-entropy strategy (shannon) or Wilcoxon test (wilcox) implemented by the Seurat package, or expression fold changes (logFC). Default: shannon.}

\item{group.size}{Specify the marker size for each subset derived from single-cell data. Default: 30.}

\item{knn.spots}{Number of nearest neighbors to consider for spot-based deconvolution analysis. When set to 0, neighbors are not used. Default: 5.}

\item{resolution}{Specify the resolution for spatial clustering. Default: 0.8.}

\item{max.cells.in.spot}{Maximum number of cells in ST spots. Default: 10 (for 10x Visium). For high-resolution ST data (such as Image-based ST technology), set \code{max.cells.in.spot} to 1.}

\item{fix.cells.in.spot}{Fixed number of cells assigned to a spot or not. Default: FALSE.}

\item{signature.scoring.method}{Method for scoring the signature of cell types in ST data: AddModuleScore, UCell, or AverageExpr. Default: AddModuleScore.}

\item{hotspot.detection.threshold}{P-value threshold for determining hot spots of cell types, range from 0 to 1. Default: 1.}

\item{adjust.deconv}{Specify the strategy for adjusting deconvolution estimation, indicating whether to incorporate hotspot weights into cellular composition estimation; if enabled, a cleaner spatial architecture may be obtained, with regions corresponding to cell types with low confidence potentially removed. Options: "Scales" (rescales hotspot significance within tissue identities), "NONE" (applies no weighting). Default: "Scales".}

\item{feature.based}{Specify whether features for likelihood or correlation calculations between single cells and spots are based on gene expression ('gene.based') or signature scores of cell types ('celltype.based'). Default: gene.based.}

\item{dist.based}{Dimensionality reduction basis used for distance weighting, UMAP or TSNE. Default: UMAP.}

\item{spatial.weight}{Whether to apply spatial weights to the analysis. Default: FALSE.}

\item{reduction}{Dimensionality reduction technique to align spatial and single-cell data. Default: cca.}

\item{platform.res}{Resolution of the spatial platform (Low or High). Default: Low.}

\item{partition}{Split into sub-modules when mapping SC to ST spots, with 'partion' set to TRUE. Default: FALSE.}

\item{n.workers}{Number of cores for parallel processing. Default: 4.}

\item{verbose}{Show running messages or not. Default: TRUE.}
}
\value{
Assigned results wrapped as a Seurat object.
}
\description{
Accurate mapping of single cells to spatial coordinates.
}
\examples{
sp.obj <- system.file("data", "Kindney_SP.RDS", package = "Cell2Spatial") \%>\% readRDS(.)
sc.obj <- system.file("data", "Kindney_SC.RDS", package = "Cell2Spatial") \%>\% readRDS(.)
sce <- runCell2Spatial(sp.obj, sc.obj, cell.type.column = "mainCtype", group.size = 30, fix.cells.in.spot = TRUE, partition = TRUE)
}
