% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{selectByProb}
\alias{selectByProb}
\title{selectByProb}
\usage{
selectByProb(sp.score, sc.score, adj.w)
}
\value{
Similarity matrix.
}
\description{
Measuring the similarity distances using probability model.
}
