% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{selectMakers}
\alias{selectMakers}
\title{selectMakers}
\usage{
selectMakers(sc.obj, cell.type.markers, group.size, assay.type, select.markers)
}
\arguments{
\item{sc.obj}{Seurat object containing SC data.}

\item{cell.type.markers}{A list of markers for each cell type. Must be provided if SC data contains only one cell type. Default: NULL.}

\item{group.size}{The size of the subsets used to find markers. Default: 100.}

\item{assay.type}{Assay type to use for marker finding. Default: RNA.}

\item{select.markers}{Strategy for inferring marker genes.}

\item{verbose}{Boolean indicating whether to show running messages. Default: TRUE.}
}
\value{
A list of marker genes for each cell type.
}
\description{
Select marker genes for each cell type based on single-cell (SC) data.
}
