% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightDist.R
\name{weightDist}
\alias{weightDist}
\title{weightDist}
\usage{
weightDist(sc.obj, sp.obj, lamba, mc.cores = 4)
}
\arguments{
\item{sc.obj}{Seurat object of SC data.}

\item{sp.obj}{Seurat object of ST data.}

\item{lamba}{Median of cell counts for spots.}

\item{mc.cores}{Number of cores for parallel running. Default: 4}
}
\value{
A matrix of weighted distance matrix. Rows represent spot clusters and columns are cell types.
}
\description{
Weight distance between SC and ST data.
}
