% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{weightSimScore}
\alias{weightSimScore}
\title{weightSimScore}
\usage{
weightSimScore(
  out.sim,
  adj.w,
  spot.name,
  cell.names,
  hot.spts = NULL,
  p.cut = 0.01
)
}
\arguments{
\item{out.sim}{Matrix or data frame of similarity scores between spots and cell types.}

\item{adj.w}{Weight matrix specifying adjustments for each spot and cell type combination.}

\item{spot.name}{Vector or list of spot names.}

\item{cell.names}{Vector or list of cell type names.}

\item{hot.spts}{A data frame indicating hotspot presence in spots and cell types.}
}
\value{
Weighted similarity scores matrix where adjustments are applied based on provided weights and hotspot information.
}
\description{
Applies weighted adjustment to similarity scores between spots and cell types based on provided adjustment weights and hotspot information.
}
