% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{weigthNetProb}
\alias{weigthNetProb}
\title{weigthNetProb}
\usage{
weigthNetProb(netx.pred, sc.obj, sp.obj, hot.spts)
}
\arguments{
\item{netx.pred}{A probability matrix output from FNN model.}

\item{sc.obj}{Seurat object of SC data.}

\item{sp.obj}{Seurt object of ST data.}

\item{hot.spts}{Boolean value matrix of hot spots.}
}
\value{
Weighetd probability matrix.
}
\description{
Weighted network probablity using the boolean value of hot spots.
}
